refactor(engine): extract rocky-ir crate (Phase G1)
Phase G1 of the blank-slate roadmap: extract the IR types from
`rocky-core` into a dedicated `rocky-ir` crate. The IR is consumed by
every adapter, the compiler, the CLI, and the AI / LSP layers; pulling
all of `rocky-core` (with its `redb`, `tokio`, state store, dialect
traits) just to see the IR types created circular-dep risk as the
typed-IR surface grew. `rocky-ir` is a thin "data only" leaf crate so
consumers can depend on the IR vocabulary without dragging the runtime
in.

Sequenced after Typed-IR Phase 3 PR-5 (#490) deleted the `Plan` enum —
extracting `rocky-ir` while `Plan` still lived in `rocky-core` would
either pull `Plan` across the crate boundary or leave it stranded
either way. Post-Phase-3, this is a pure namespace move.

New crate layout (engine/crates/rocky-ir/):
- `lib.rs` — module root + flat re-exports
- `ir.rs` — `ModelIr`, `ProjectIr`, `ModelIrVariant`,
  `MaterializationStrategy`, `ColumnSelection`, `MetadataColumn`,
  `GovernanceConfig`, `ColumnMask`, `ColumnInfo`, `DriftAction`,
  `WatermarkState`, `PartitionWindow`, `SourceRef`, `TargetRef`,
  `TableRef`, `Permission`, `Grant`, `GrantTarget`, `ResolvedRole`
  (the verbatim contents of the previous `rocky-core/src/ir.rs`)
- `lakehouse.rs` — `LakehouseFormat`, `LakehouseOptions`,
  `LakehouseError` (data only; the dialect-aware
  `generate_lakehouse_ddl` stays in `rocky-core/src/lakehouse.rs`
  because it consumes `SqlDialect`)
- `lineage.rs` — `LineageEdge`, `QualifiedColumn`
- `types.rs` — `RockyType`, `TypedColumn`, `StructField`,
  `common_supertype`, `is_assignable`
- `dag.rs` — `DagNode`, `DagError`, `topological_sort`,
  `execution_layers`
- `time_grain.rs` — extracted from `rocky-core/src/models.rs`;
  referenced by `MaterializationStrategy::TimeInterval` and
  `Microbatch`
- `mask.rs` — `MaskStrategy` extracted from
  `rocky-core/src/traits.rs`; referenced by `ColumnMask`

`rocky-ir` deps: `rocky-sql`, `serde`, `serde_json`, `schemars`,
`chrono`, `blake3`, `thiserror`. No `tokio`, no `redb`, no
`object_store`, no `reqwest`. Leaf-crate clean.

rocky-core changes:
- Cargo.toml: add `rocky-ir = { path = "../rocky-ir" }` dep.
- `src/ir.rs` — collapsed from 1423 lines to a 6-line `pub use
  rocky_ir::*` shim. Existing `use crate::ir::X` call sites in the
  ~25 rocky-core source files keep working unchanged.
- `src/lineage.rs`, `src/types.rs`, `src/dag.rs` — same shim pattern.
- `src/lakehouse.rs` — the data types (`LakehouseFormat`,
  `LakehouseOptions`, `LakehouseError`) are re-exported from
  `rocky_ir::lakehouse`; the DDL generator
  (`generate_lakehouse_ddl`) and its dialect-specific helpers
  (Delta / Iceberg / MV / streaming-table / view emitters) stay in
  this file (they need `SqlDialect`). File shrinks from 966 lines
  to 883 lines (the ~80-line data-type block moves out).
- `src/models.rs`: delete the inline `TimeGrain` definition (~80
  lines); replace with `pub use rocky_ir::TimeGrain` re-export so
  existing `crate::models::TimeGrain` paths keep working.
- `src/traits.rs`: delete the inline `MaskStrategy` definition (~30
  lines); replace with `pub use rocky_ir::MaskStrategy`. Drop
  now-unused `schemars::JsonSchema` import.

External consumer crates: NO CHANGES this PR. All ~67 cross-crate
`use rocky_core::ir::X` paths keep working via the rocky-core
re-export shim. A follow-up PR (G1b) will sweep them to
`use rocky_ir::X` directly and delete the rocky-core shim modules.

Net diff: ~145 lines deleted from existing rocky-core files (data
types replaced with re-exports), plus new rocky-ir crate (~2,800
lines moved verbatim).

Why split rather than full sweep in one PR:
- The shim approach lands a clean dep-graph cut (rocky-core →
  rocky-ir, no reverse edge) without touching 8 downstream crates.
- The rocky-ir surface area exposed in the PR is exactly what was
  in `rocky-core/src/ir.rs` plus the items it referenced. No
  semantic changes.
- The follow-up sweep PR is purely mechanical (sed-style rename of
  `rocky_core::ir::` → `rocky_ir::`) and easier to review against
  this as the baseline.

Verification:
- `cargo build --workspace`: clean
- `cargo test -p rocky-ir --lib`: 53 passed (the IR-test moved with
  the extraction)
- `cargo test -p rocky-core --lib`: 1141 passed (1194 - 53 that
  moved to rocky-ir)
- `cargo test -p rocky-core --test e2e`: 30 passed
- `cargo test -p rocky-core --test integration`: 20 passed
- `cargo test -p rocky-cli --lib`: 376 passed
- `cargo test -p rocky-cli --test ir_golden`: 4 passed (snapshot
  byte-stability + recipe-hash pin)
- `cargo clippy --workspace --all-targets -- -D warnings`: clean
- `cargo fmt --check`: clean
- `examples/playground/pocs/00-foundations/00-playground-default/run.sh`:
  green

`ir_golden` byte-stability is the key invariant: `ir.json` and the
per-dialect SQL fixtures are unchanged, confirming that the IR types
moved across the crate boundary without drift in serialization shape
or recipe-hash inputs.
