+++ engine/Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 members = [
     "crates/rocky-core",
+    "crates/rocky-ir",
     "crates/rocky-sql",
     "crates/rocky-compiler",
     "crates/rocky-lang",
