+++ engine/crates/rocky-core/Cargo.toml
@@ -7,6 +7,7 @@ license.workspace = true
 rust-version.workspace = true
 
 [dependencies]
+rocky-ir = { path = "../rocky-ir" }
 rocky-sql = { path = "../rocky-sql" }
 async-trait = { workspace = true }
 serde = { workspace = true }
